support extend command in plugin
