+++ packages/umi-build-dev/src/getPlugins.js
@@ -57,6 +57,7 @@ Try:
 
   // 内置插件
   const builtInPlugins = [
+    './plugins/commands/dev',
     './plugins/output-path',
     './plugins/global-js',
     './plugins/global-css',
