+++ packages/umi-preset-react/src/plugins/dynamicImport.js
