+++ packages/umi-preset-react/src/plugins/hardSource.js
