+++ packages/umi-preset-react/src/plugins/loadingComponent.js
