+++ packages/umi-preset-react/src/plugins/mobile/hd/template/flex.js
