+++ packages/umi-preset-react/src/plugins/mobile/hd/template/index.js
