+++ packages/umi-utils/src/compatDirname.js
@@ -1,7 +1,7 @@
 import { dirname, join } from 'path';
 import { existsSync } from 'fs';
 
-export default function(path, cwd) {
+export default function(path, cwd, fallback) {
   const pkgPath = join(cwd, 'package.json');
   const library = path.split('/')[0];
   if (existsSync(pkgPath)) {
@@ -10,5 +10,5 @@ export default function(path, cwd) {
       return dirname(join(cwd, 'node_modules', path));
     }
   }
-  return dirname(require.resolve(path));
+  return fallback;
 }
