+++ packages/umi-utils/src/index.js
@@ -1,0 +1,3 @@
+export { default as findJS } from './findJS';
+export { default as compatDirname } from './compatDirname';
+export { default as winPath } from './winPath';
