+++ packages/umi-utils/src/winPath.js
@@ -1,0 +1,5 @@
+import slash from 'slash2';
+
+export default function(path) {
+  return slash(path);
+}
