+++ packages/umi/src/buildDevOpts.js
@@ -1,4 +1,4 @@
-import { resolve, join } from 'path';
+import { resolve, join, dirname } from 'path';
 import isAbsolute from 'path-is-absolute';
 import isWindows from 'is-windows';
 import slash from 'slash2';
@@ -6,7 +6,6 @@ import slash from 'slash2';
 const debug = require('debug')('umi:devDevOpts');
 
 export default function(opts = {}) {
-  const { extraResolveModules, hash } = opts;
   debug(`opts: ${JSON.stringify(opts)}`);
   delete opts.extraResolveModules;
 
@@ -22,25 +21,9 @@ export default function(opts = {}) {
     }
   }
 
+  process.env.UMI_DIR = dirname(require.resolve('../package'));
+
   return {
     cwd,
-    // eslint-disable-line
-    babel: resolve(__dirname, './babel'),
-    extraResolveModules: [
-      ...(extraResolveModules || []),
-      resolve(__dirname, '../../node_modules'),
-    ],
-    libraryAlias: {
-      dynamic: require.resolve('./dynamic'),
-      link: require.resolve('./link'),
-      navlink: require.resolve('./navlink'),
-      redirect: require.resolve('./redirect'),
-      router: require.resolve('./router'),
-      withRouter: require.resolve('./withRouter'),
-      _renderRoutes: require.resolve('./renderRoutes'),
-      _createHistory: require.resolve('./createHistory'),
-    },
-    hash,
-    ...opts,
   };
 }
