+++ packages/umi/src/scripts/realDev.js
@@ -10,7 +10,4 @@ process.on('SIGINT', () => {
 
 dev({
   plugins: argv.plugins ? argv.plugins.split(',') : [],
-}).catch(e => {
-  console.log(e);
-  process.exit(1);
 });
